requirements.txt
pyodbc
fastapi
uvicorn[standard]
pydantic
azure-identity
paho-mqtt
mysql-connector-python